backend/pyproject.toml
[project]
name = "werbesystem-backend"
version = "0.1.0"
description = "Interactive LLM-powered advertising system backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.20.0",
    "alembic>=1.14.0",
    "chromadb>=0.6.0",
    "langgraph>=0.3.0",
    "langgraph-checkpoint-sqlite>=2.0.0",
    "langchain-openai>=0.3.0",
    "httpx>=0.28.0",
    "beautifulsoup4>=4.12.0",
    "duckduckgo-search>=7.0.0",
    "pydantic-settings>=2.7.0",
    "pyyaml>=6.0",
    "sentence-transformers>=3.4.0",
    "websockets>=14.0",
    "playwright>=1.50.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"
